requirements.txt
langchain == 0.2.1
python-decouple==3.8
langchain-google-genai==1.0.5
langchain-openai==0.1.7
streamlit==1.35.0
wikipedia==1.4.0
numexpr==2.10.0
duckduckgo-search==6.1.2
chromadb==0.4.18
pypdf==4.2.0
docx2txt==0.8
youtube-transcript-api==0.6.2
